requirements.txt
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4
sqlalchemy==2.0.23
python-dateutil==2.8.2
alembic==1.12.0